fix(Catalog): Workaround for string-defined default values
In Camel 4.15, the `default` values are written in `string`, regardless of its type. For instance, this is how the `scriptPropertyPlaceholders` property is defined, notice the `default: "true"` portion.

```
          "scriptPropertyPlaceholders" : {
            "type" : "boolean",
            "title" : "Script Property Placeholders",
            "description" : "Whether the script should support using Camel property placeholder syntax {{ }}.",
            "default" : "true"
          },
```

This commit fixes this issue at the Entity and Pattern level by
adjusting the default value to its right type.

A fix for this was provided in Camel upstream here: https://github.com/apache/camel/pull/19753